cmd/cmart-pwd: initial commit, porting over from shell script
Initial testing indicates that this takes about 0.44s, as shown
through the following shell execution:

    for i in {1..10}; do time smart-pwd; done

Testing the function linked in the comments in a similar manner
yields runtimes of about 0.065s. These ratios back up the understanding
that most of the time spent here is in forking the subprocesses.

Thus a key area of optimization here would be to eliminate the
git calls entirely. The current win is likely entirely due to the
elimination of the perl command from the original script.